Update using LiveData
